requirements.txt
# COGNITUS - Python Dependencies
# For NVIDIA Jetson Orin Nano with Python 3.8 and ROS2 Foxy

# Deep Learning Frameworks
# PyTorch installed separately via Jetson-specific wheel
transformers==4.35.0
accelerate==0.25.0
bitsandbytes==0.41.0
optimum==1.14.0

# Computer Vision
opencv-python==4.5.5.64
ultralytics==8.0.200
Pillow==9.5.0

# Speech Processing
openai-whisper==20231117
piper-tts==1.2.0
pyaudio==0.2.13
soundfile==0.12.1
librosa==0.10.1

# NLP and Embeddings
sentence-transformers==2.2.2
tokenizers==0.15.0

# Vector Database
chromadb==0.4.18
duckdb==0.9.2

# ROS2 Integration
rclpy  # Installed via apt
cv-bridge  # Installed via apt
sensor-msgs  # Installed via apt

# Utilities
numpy==1.23.5
scipy==1.10.1
pandas==2.0.3
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
psutil==5.9.6

# Model Optimization
onnx==1.15.0
onnxruntime-gpu==1.16.3
tensorrt  # Installed with JetPack

# Audio Processing
webrtcvad==2.0.10
noisereduce==3.0.0

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
